requirements.txt
bidict==0.22.0
castepxbin==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
colormath==3.0.0
contourpy==1.0.5
cycler==0.11.0
emmet-core==0.38.5
fonttools==4.38.0
future==0.18.2
h11==0.14.0
h5py==3.7.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
latexcodec==2.0.1
matplotlib==3.6.1
monty==2022.9.9
mp-api==0.29.3
mpmath==1.2.1
msgpack==1.0.4
networkx==2.8.7
numpy==1.23.4
packaging==21.3
palettable==3.3.0
pandas==1.5.1
pathlib==1.0.1
phonopy==2.16.3
Pillow==9.2.0
plotly==5.10.0
pybtex==0.24.0
pydantic==1.10.2
pymatgen==2022.10.22
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schedule==1.1.0
scikit-learn==1.1.2
scipy==1.9.3
seekpath==2.0.1
six==1.16.0
sklearn==0.0
sumolib==1.14.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
traci==1.14.1
typing_extensions==4.4.0
uncertainties==3.1.7
urllib3==1.26.12
uvicorn==0.19.0
websockets==10.3
wincertstore==0.2
